Fix ios warning "values of type 'OSStatus' should not be used as format arguments; add an explicit cast to 'int' instead [-Wformat]"

+++ ios/RNLiveAudioStream.m
@@ -47,7 +47,7 @@ RCT_EXPORT_METHOD(start) {
 
     OSStatus status = AudioQueueNewInput(&_recordState.mDataFormat, HandleInputBuffer, &_recordState, NULL, NULL, 0, &_recordState.mQueue);
     if (status != 0) {
-        RCTLog(@"[RNLiveAudioStream] Record Failed. Cannot initialize AudioQueueNewInput. status: %i", status);
+        RCTLog(@"[RNLiveAudioStream] Record Failed. Cannot initialize AudioQueueNewInput. status: %i", (int) status);
         return;
     }
